# 8287835: Add support for additional float/double to integral conversion for x86