--- a/src/hotspot/cpu/x86/x86.ad
+++ b/src/hotspot/cpu/x86/x86.ad
@@ -1868,12 +1868,15 @@ const bool Matcher::match_rule_supported_vector(int opcode, int vlen, BasicType 
       }
       break;
     case Op_VectorCastD2X:
-      if (((UseAVX <= 2) || (!VM_Version::supports_avx512vl())) &&
-          (is_subword_type(bt) || bt == T_INT)) {
-        return false;
-      }
-      if (is_integral_type(bt) && !VM_Version::supports_avx512dq()) {
-        return false;
+      // Conversion to integral type is only supported on AVX-512 platforms with avx512dq.
+      // Need avx512vl for size_in_bits < 512
+      if (is_integral_type(bt)) {
+        if (!VM_Version::supports_avx512dq()) {
+          return false;
+        }
+        if (size_in_bits < 512 && !VM_Version::supports_avx512vl()) {
+          return false;
+        }
       }
       break;
     case Op_RoundVD:
@@ -1882,12 +1885,20 @@ const bool Matcher::match_rule_supported_vector(int opcode, int vlen, BasicType 
       }
       break;
     case Op_VectorCastF2X:
-      if (((UseAVX <= 2) || (!VM_Version::supports_avx512vl())) &&
-          (is_subword_type(bt) || bt == T_LONG)) {
-        return false;
-      }
-      if ((bt == T_LONG) && !VM_Version::supports_avx512dq()) {
-        return false;
+      // F2I is supported on all AVX and above platforms
+      // For conversion to other integral types need AVX512:
+      //     Conversion to long in addition needs avx512dq
+      //     Need avx512vl for size_in_bits < 512
+      if (is_integral_type(bt) && (bt != T_INT)) {
+        if (UseAVX <= 2) {
+          return false;
+        }
+        if ((bt == T_LONG) && !VM_Version::supports_avx512dq()) {
+          return false;
+        }
+        if (size_in_bits < 512 && !VM_Version::supports_avx512vl()) {
+          return false;
+        }
       }
       break;
     case Op_MulReductionVI:
@@ -7277,6 +7288,8 @@ instruct vcastFtoD_reg(vec dst, vec src) %{
 
 
 instruct castFtoI_reg_avx(vec dst, vec src, vec xtmp1, vec xtmp2, vec xtmp3, vec xtmp4, rRegP scratch, rFlagsReg cr) %{
+  // F2I conversion for < 64 byte vector using AVX instructions
+  // AVX512 platforms that dont support avx512vl also use AVX instructions to support F2I
   predicate(!VM_Version::supports_avx512vl() &&
             Matcher::vector_length_in_bytes(n) < 64 &&
             Matcher::vector_element_basic_type(n) == T_INT);
@@ -7292,10 +7305,29 @@ instruct castFtoI_reg_avx(vec dst, vec src, vec xtmp1, vec xtmp2, vec xtmp3, vec
   ins_pipe( pipe_slow );
 %}
 
+instruct castFtoI_reg_evex(vec dst, vec src, vec xtmp1, vec xtmp2, kReg ktmp1, kReg ktmp2, rRegP scratch, rFlagsReg cr) %{
+  predicate((VM_Version::supports_avx512vl() ||
+             Matcher::vector_length_in_bytes(n) == 64) &&
+             Matcher::vector_element_basic_type(n) == T_INT);
+  match(Set dst (VectorCastF2X src));
+  effect(TEMP dst, TEMP xtmp1, TEMP xtmp2, TEMP ktmp1, TEMP ktmp2, TEMP scratch, KILL cr);
+  format %{ "vector_cast_f2i $dst,$src\t! using $xtmp1, $xtmp2, $ktmp1, $ktmp2 and $scratch as TEMP" %}
+  ins_encode %{
+    int vlen_enc = vector_length_encoding(this);
+    __ vector_castF2I_evex($dst$$XMMRegister, $src$$XMMRegister, $xtmp1$$XMMRegister,
+                           $xtmp2$$XMMRegister, $ktmp1$$KRegister, $ktmp2$$KRegister,
+                           ExternalAddress(vector_float_signflip()), $scratch$$Register, vlen_enc);
+  %}
+  ins_pipe( pipe_slow );
+%}
+
 instruct castFtoX_reg_evex(vec dst, vec src, vec xtmp1, vec xtmp2, kReg ktmp1, kReg ktmp2, rRegP scratch, rFlagsReg cr) %{
-  predicate(((VM_Version::supports_avx512vl() ||
-              Matcher::vector_length_in_bytes(n) == 64)) &&
-             is_integral_type(Matcher::vector_element_basic_type(n)));
+  // F2X conversion for integral non T_INT target using AVX512 instructions
+  // Platforms that dont support avx512vl can only support 64 byte vectors
+  predicate((VM_Version::supports_avx512vl() ||
+             Matcher::vector_length_in_bytes(n) == 64) &&
+            is_integral_type(Matcher::vector_element_basic_type(n)) &&
+            Matcher::vector_element_basic_type(n) != T_INT);
   match(Set dst (VectorCastF2X src));
   effect(TEMP dst, TEMP xtmp1, TEMP xtmp2, TEMP ktmp1, TEMP ktmp2, TEMP scratch, KILL cr);
   format %{ "vector_cast_f2x $dst,$src\t! using $xtmp1, $xtmp2, $ktmp1, $ktmp2 and $scratch as TEMP" %}
@@ -7313,7 +7345,8 @@ instruct castFtoX_reg_evex(vec dst, vec src, vec xtmp1, vec xtmp2, kReg ktmp1, k
                              ExternalAddress(vector_float_signflip()), $scratch$$Register, vlen_enc);
       if (to_elem_bt == T_SHORT) {
         __ evpmovdw($dst$$XMMRegister, $dst$$XMMRegister, vlen_enc);
-      } else if (to_elem_bt == T_BYTE){
+      } else {
+        assert(to_elem_bt == T_BYTE, "required");
         __ evpmovdb($dst$$XMMRegister, $dst$$XMMRegister, vlen_enc);
       }
     }